requirements.txt
django 
djangorestframework 
google-generativeai 
pypdf 
chromadb 
langchain-community 
tiktoken